+++ installer/src/platforms/linux/mod.rs
@@ -95,8 +95,8 @@ pub fn check_build_environment() {
     if !rustup_toolchains_string.contains("stable") {
         eprintln!("WhiteBeam: No stable Rust found in toolchain, consider running: rustup toolchain install stable");
         std::process::exit(1);
-    } else */ if !(rustup_toolchains_string.contains("nightly-2022-03-01")) {
-        eprintln!("WhiteBeam: No pinned nightly Rust found in toolchain, consider running: rustup toolchain install nightly-2022-03-01");
+    } else */ if !(rustup_toolchains_string.contains("nightly-2022-06-25")) {
+        eprintln!("WhiteBeam: No pinned nightly Rust found in toolchain, consider running: rustup toolchain install nightly-2022-06-25");
         std::process::exit(1);
     }
 }